+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added missing `pageSize` and `page` parameters to `TopHeadlines` endpoint.
+- Added missing `excludeDomains` parameter to `Everything` endpoint.
 
 ### Deprecated
 - Removed *language* parameter from `TopHeadlines` endpoint. It didn't work in the first place, so this shouldn't affect anything.